[Refactoring] Change Restaurants files
- Change folder to restaurants
- Create Restaurants component
