+++ src/containers/restaurants/CategoriesContainer.test.jsx
